requirements.txt
pandas
psycopg2-binary
dash
plotly